# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TechZone Electronics & Gaming Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers/register:
    post:
      operationId: customerRegister
      tags: [customers]
      summary: Register a new customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/login:
    post:
      operationId: customerLogin
      tags: [customers]
      summary: Customer login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
...
  /customers/me:
    get:
      operationId: customerMe
      tags: [customers]
      summary: Current customer
      responses:
        "200":
          description: Current customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "401":
          description: Not authenticated
    put:
      operationId: customerUpdateProfile
      tags: [customers]
      summary: Update current customer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
...
  /admin/gift-cards:
    get:
      operationId: adminListGiftCards
      tags: [admin-gift-cards]
      summary: List gift cards
      responses:
        "200":
          description: Gift cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GiftCard"
    post:
      operationId: adminCreateGiftCard
      tags: [admin-gift-cards]
      summary: Create a gift card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftCardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCard"
  /admin/gift-cards/{id}:
    put:
      operationId: adminUpdateGiftCard
      tags: [admin-gift-cards]
      summary: Update a gift card
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftCardUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCard"
        "404":
          description: Not found
    delete:
      operationId: adminDeleteGiftCard
      tags: [admin-gift-cards]
      summary: Delete a gift card
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/subscription-plans:
    get:
      operationId: adminListSubscriptionPlans
      tags: [admin-subscription-plans]
      summary: List subscription plans
      responses:
        "200":
          description: Subscription plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubscriptionPlan"
    post:
      operationId: adminCreateSubscriptionPlan
      tags: [admin-subscription-plans]
      summary: Create a subscription plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionPlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPlan"
  /admin/subscription-plans/{id}:
    put:
      operationId: adminUpdateSubscriptionPlan
      tags: [admin-subscription-plans]
      summary: Update a subscription plan
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionPlanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPlan"
        "404":
          description: Not found
    delete:
      operationId: adminDeleteSubscriptionPlan
      tags: [admin-subscription-plans]
      summary: Delete a subscription plan
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /gift-cards/redeem:
    post:
      operationId: redeemGiftCard
      tags: [gift-cards]
      summary: Redeem a gift card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftCardRedeemInput"
      responses:
        "200":
          description: Redeem result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCardRedeemResult"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: ["string", "null"] }
...
    Customer:
      type: object
      required: [id, email, fullName, isActive, createdAt]
      properties:
        id:          { type: string }
        email:       { type: string }
        fullName:    { type: string }
        phone:       { type: ["string", "null"] }
        city:        { type: ["string", "null"] }
        address:     { type: ["string", "null"] }
        isActive:    { type: boolean }
        lastLoginAt: { type: ["string", "null"] }
        createdAt:   { type: string }
    CustomerRegisterInput:
      type: object
      required: [email, fullName, password]
      properties:
        email:    { type: string }
        fullName: { type: string }
        phone:    { type: ["string", "null"] }
        password: { type: string, minLength: 8 }
    CustomerLoginInput:
      type: object
      required: [email, password]
      properties:
        email:    { type: string }
        password: { type: string }
    CustomerProfileUpdate:
      type: object
      properties:
        fullName: { type: string }
        phone:    { type: ["string", "null"] }
        city:     { type: ["string", "null"] }
        address:  { type: ["string", "null"] }
    GiftCard:
      type: object
      required: [id, code, amount, balance, status, createdAt]
      properties:
        id:          { type: string }
        code:        { type: string }
        amount:      { type: number }
        balance:     { type: number }
        status:      { type: string, enum: [active, redeemed, expired, cancelled] }
        customerId:  { type: ["string", "null"] }
        redeemedAt:  { type: ["string", "null"] }
        expiresAt:   { type: ["string", "null"] }
        createdAt:   { type: string }
    GiftCardInput:
      type: object
      required: [code, amount]
      properties:
        id:          { type: string }
        code:        { type: string }
        amount:      { type: number }
        balance:     { type: number }
        status:      { type: string, enum: [active, redeemed, expired, cancelled] }
        expiresAt:   { type: ["string", "null"] }
    GiftCardUpdate:
      type: object
      properties:
        code:        { type: string }
        amount:      { type: number }
        balance:     { type: number }
        status:      { type: string, enum: [active, redeemed, expired, cancelled] }
        expiresAt:   { type: ["string", "null"] }
    GiftCardRedeemInput:
      type: object
      required: [code]
      properties:
        code: { type: string }
    GiftCardRedeemResult:
      type: object
      required: [success, balance]
      properties:
        success:   { type: boolean }
        balance:   { type: number }
        amount:    { type: number }
        message:   { type: ["string", "null"] }
    SubscriptionPlan:
      type: object
      required: [id, nameAr, nameEn, slug, price, period, isActive, createdAt]
      properties:
        id:            { type: string }
        nameAr:        { type: string }
        nameEn:        { type: string }
        slug:          { type: string }
        descriptionAr: { type: ["string", "null"] }
        descriptionEn: { type: ["string", "null"] }
        price:         { type: number }
        period:        { type: string, enum: [monthly, quarterly, yearly] }
        features:      { type: array, items: { type: string } }
        isActive:      { type: boolean }
        createdAt:     { type: string }
    SubscriptionPlanInput:
      type: object
      required: [nameAr, nameEn, slug, price, period]
      properties:
        id:            { type: string }
        nameAr:        { type: string }
        nameEn:        { type: string }
        slug:          { type: string }
        descriptionAr: { type: ["string", "null"] }
        descriptionEn: { type: ["string", "null"] }
        price:         { type: number }
        period:        { type: string, enum: [monthly, quarterly, yearly] }
        features:      { type: array, items: { type: string } }
        isActive:      { type: boolean }
    SubscriptionPlanUpdate:
      type: object
      properties:
        nameAr:        { type: string }
        nameEn:        { type: string }
        slug:          { type: string }
        descriptionAr: { type: ["string", "null"] }
        descriptionEn: { type: ["string", "null"] }
        price:         { type: number }
        period:        { type: string, enum: [monthly, quarterly, yearly] }
        features:      { type: array, items: { type: string } }
        isActive:      { type: boolean }
...