# lib/api-zod/src/generated/types/giftCardInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TechZone Electronics & Gaming Store API
 * OpenAPI spec version: 0.1.0
 */
import type { GiftCardInputStatus } from './giftCardInputStatus';

export interface GiftCardInput {
  id?: string;
  code: string;
  amount: number;
  balance?: number;
  status?: GiftCardInputStatus;
  /** @nullable */
  expiresAt?: string | null;
}