# lib/api-zod/src/generated/types/giftCardRedeemResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TechZone Electronics & Gaming Store API
 * OpenAPI spec version: 0.1.0
 */

export interface GiftCardRedeemResult {
  success: boolean;
  balance: number;
  amount?: number;
  /** @nullable */
  message?: string | null;
}