# lib/api-zod/src/generated/types/subscriptionPlanInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TechZone Electronics & Gaming Store API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionPlanInputPeriod } from './subscriptionPlanInputPeriod';

export interface SubscriptionPlanInput {
  id?: string;
  nameAr: string;
  nameEn: string;
  slug: string;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  descriptionEn?: string | null;
  price: number;
  period: SubscriptionPlanInputPeriod;
  features?: string[];
  isActive?: boolean;
}