# lib/api-zod/src/generated/types/subscriptionPlanUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TechZone Electronics & Gaming Store API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionPlanUpdatePeriod } from './subscriptionPlanUpdatePeriod';

export interface SubscriptionPlanUpdate {
  nameAr?: string;
  nameEn?: string;
  slug?: string;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  descriptionEn?: string | null;
  price?: number;
  period?: SubscriptionPlanUpdatePeriod;
  features?: string[];
  isActive?: boolean;
}